pyproject.toml
[project]
name = "custom-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai-agents>=0.8.4",
    "python-dotenv>=1.2.1",
]
